Dzil'ify and document
Stub dist.ini. Building the module does not yet work because I can't
find a way to have libpropcalc built and be available to the module.
FFI::Build apparently eschews `make` and is made for very simple
libraries(?), and I could not get Alien::Build to work all the way
to the end.

I would not like to provide libprocalc as an Alien module of its own
right now, because I develop the C++ library and the Perl module
at the same time, linking commits here to commits in the library.
It would be more convenient to bundle the C++ sources with the Perl
module as a git submodule or copies in a .tar.gz. There is a manual
Makefile for me while I develop now. Will worry about the proper
procedure when a release is nearing.

Also bump to libpropcalc @92b4641 which brings constants and bracket-
less symbolics (mentioned in the documentation) but no changes to the
wrapper code.